backend/requirements.txt
beautifulsoup4
fastapi
uvicorn[standard]  # Enables nice features like auto-reload
pydantic
requests
streamlit
pydantic[email]
google-api-python-client==2.108.0
openai
python-dateutil